Refactor middleware utils to separate file
